requirements.txt
streamlit==1.28.1
openai-whisper==20231117
torch>=2.1.0
transformers>=4.53.0
tokenizers>=0.21.0
datasets==2.14.6
spacy==3.7.2
nltk==3.8.1
moviepy==1.0.3
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.7
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.0
ffmpeg-python==0.2.0
pyaudio>=0.2.11
sounddevice>=0.4.6
plotly>=5.0.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
flask-sock==0.7.0
simple-websocket==1.0.0
